Cleanup of cli: separate cli module
