+++ rumbas/src/cli/mod.rs
@@ -1,0 +1,10 @@
+mod compile;
+mod import;
+mod init;
+pub mod logger;
+mod schema;
+
+pub use compile::compile;
+pub use import::import;
+pub use init::init;
+pub use schema::schema;
